recomputed ec/Q tables for 140000 <= rad2N <= 160000
In this range of radicals of 2N (in particular around 148500 - 150500), I made a systematic error in the original computation, and thus I had to recompute the tables.
To explain the origin of this systematic error, first we sketch the general strategy of the computation:
For a given rad2N = prod(S), first the list of potential two-division fields is computed (and cached), then for each such two-division field a certain constraint S-unit equation is solved, and from its solutions the elliptic curves with good reduction outside S with the given two-division field is obtained.
Now, in this above range of rad2N, several CPUs were working on the same set S asynchronously.
The problem with that is that the order of the computed lists of two-division fields for a fixed S is not always the same, as the computation of degree 3 extensions included some (unforseen) randomness.
Thus, in case two processes started before a list of two-division fields was cached, the processes worked with these different lists and the full set of j-invariants for some given S was collected in an erroneous way.

	modified:   elliptic-curve-tables/bounded-rad2N/K_1.1.1.1*/js_K_1.1.1.1_rad2N_140000-150000_sobjs*.zip
	modified:   elliptic-curve-tables/bounded-rad2N/K_1.1.1.1*/js_K_1.1.1.1_rad2N_140000-150000_txts*.zip
	modified:   elliptic-curve-tables/bounded-rad2N/K_1.1.1.1*/js_K_1.1.1.1_rad2N_150000-160000_sobjs*.zip
	modified:   elliptic-curve-tables/bounded-rad2N/K_1.1.1.1*/js_K_1.1.1.1_rad2N_150000-160000_txts*.zip